Updated Post, tag and user entities.

+++ src/main/java/com/example/demo/DAL/Interfaces/TagsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface TagsRepository extends JpaRepository<Tag, Long> {
+public interface TagsRepository extends JpaRepository<Tag, Integer> {
 
     @Query("DELETE FROM Tag t " +
             "WHERE t.id in (:tags)")
